Create a dust model from a C++ input file

Create a dust model from a C++ input file. This function will compile the dust support around your model and return an object that can be used to work with the model (see the Details below).

dust(
  filename,
  type = NULL,
  name = NULL,
  quiet = FALSE,
  workdir = NULL,
  gpu = FALSE
)

Arguments

filename
The path to a single C++ file

type
The name of the "type" (the C++ class) that represents your model. If NULL we try to work this out from your file using extremely simple heuristics.

name
The name of the model; for technical reasons this must be alphanumeric characters only (sorry, no underscore) and must not start with a number. If NULL the value of type will be used.

quiet
Logical, indicating if compilation messages from pkgbuild should be displayed. Error messages will be displayed on compilation failure regardless of the value used.

workdir
Optional working directory to use. If NULL uses a temporary directory. By using a different directory of your choosing you can see the generated code.

gpu
Create GPU code

Input requirements

Your input dust model must satisfy a few requirements.

Define some class that implements your model (below model is assumed to be the class name)

That class must define a type init_t (so model::init_t) that contains its internal data and the model must be constructable with a const reference to this type (const model::init_t& data)

That class must also include a typedef that describes the model's floating point type, real_t. Most models can include typedef double real_t; in their public section.

The model must have a method size() returning size_t which returns the size of the system. This size may depend on values in your initialisation object but is constant within a model run.

The model must have a method initial (which may not be const), taking a step number (size_t) and returning a std::vector<real_t> of initial state for the model.

The model must have a method update taking arguments:

size_t step: the step number

const double * state: the state at the beginning of the step

dust::rng_state_t<real_t>& rng_state: the dust random number generator state - this must be a reference, as it will be modified as random numbers are drawn

double *state_next: the end state of the model (to be written to by your function)

Your update function is the core here and should update the state of the system - you are expected to update all values of state on return.

It is very important that none of the functions in the class use the R API in any way as these functions will be called in parallel.

You must also provide a data-wrangling function for producing an object of type model::init_t from an R list. We use cpp11 for this. Your function will look like:

template &lt;&gt;
model::init_t dust_data&lt;model&gt;(cpp11::list data) {
  return ...;
}

With the body interacting with data to create an object of type model::init_t and returning it. This function will be called in serial and may use anything in the cpp11 API. All elements of the returned object must be standard C/C++ (e.g., STL) types and not cpp11/R types.

Your model may provided a template specialisation dust_data<model::init_t>() returning a cpp11::sexp for returning arbitrary information back to the R session:

template &lt;&gt;
cpp11::sexp dust_info&lt;model&gt;(const model::init_t&amp; data) {
  return cpp11::wrap(...);
}

What you do with this is up to you. If not present then the info() method on the created object will return NULL. Potential use cases for this are to return information about variable ordering, or any processing done while accepting the data object used to create the data fed into the particles.

Methods

Public methods

dust_class$new()

dust_class$name()

dust_class$run()

dust_class$set_index()

dust_class$index()

dust_class$n_threads()

dust_class$set_state()

dust_class$reset()

dust_class$state()

dust_class$step()

dust_class$reorder()

dust_class$info()

dust_class$data()

dust_class$rng_state()

dust_class$has_openmp()

Method new()

Create a new model

Usage

dust_class$new(data, step, n_particles, n_threads = 1L, seed = NULL)

Arguments

data

Data to initialise your model with; a list object, but the required elements will depend on the details of your model.

step

Initial step - must be nonnegative

n_particles

Number of particles to create - must be at least 1

n_threads

Number of OMP threads to use, if dust and your model were compiled with OMP support (details to come). n_particles should be a multiple of n_threads (e.g., if you use 8 threads, then you should have 8, 16, 24, etc particles). However, this is not compulsary.

seed

The seed to use for the random number generator. Can be a positive integer, NULL (initialise with R's random number generator) or a raw vector of a length that is a multiple of 32 to directly initialise the generator (e..g., from the dust object's $rng_state() method).

Method name()

Returns friendly model name

Usage

dust_class$name()

Method run()

Run the model up to a point in time, returning the filtered state at that point.

Usage

dust_class$run(step_end)

Arguments

step_end

Step to run to (if less than or equal to the current step(), silently nothing will happen)

Method set_index()

Set the "index" vector that is used to return a subset of data after using run(). If this is not used then run() returns all elements in your state vector, which may be excessive and slower than necessary. This method must be called after any call to reset() as reset() may change the size of the state and that will invalidate the index.

Usage

dust_class$set_index(index)

Arguments

index

The index vector - must be an integer vector with elements between 1 and the length of the state (this will be validated, and an error thrown if an invalid index is given).

Method index()

Returns the index as set by $set_index

Usage

dust_class$index()

Method n_threads()

Returns the number of threads that the model was constructed with

Usage

dust_class$n_threads()

Method set_state()

Set the "state" vector for all particles, overriding whatever your models initial() method provides.

Usage

dust_class$set_state(state, step = NULL)

Arguments

state

The state vector - can be either a numeric vector with the same length as the model's current state (in which case the same state is applied to all particles), or a numeric matrix with as many rows as your model's state and as many columns as you have particles (in which case you can set a number of different starting states at once).

step

If not NULL, then this sets the initial step. If this is a vector (with the same length as the number of particles), then particles are started from different initial steps and run up to the larges step given (i.e., max(step))

Method reset()

Reset the model while preserving the random number stream state

Usage

dust_class$reset(data, step)

Arguments

data

New data for the model (see constructor)

step

New initial step for the model (see constructor)

Method state()

Return full model state

Usage

dust_class$state(index = NULL)

Arguments

index

Optional index to select state using

Method step()

Return current model step

Usage

dust_class$step()

Method reorder()

Reorder or resample particles.

Usage

dust_class$reorder(index)

Arguments

index

An integer vector, with values between 1 and n_particles, indicating the index of the current particles that new particles should take.

Method info()

Returns information about the data that your model was created with. Only returns non-NULL if the model provides a dust_info template specialisation.

Usage

dust_class$info()

Method data()

Returns the data object that your model was constructed with.

Usage

dust_class$data()

Method rng_state()

Returns the state of the random number generator. This returns a raw vector of length 32 * n_particles. This can be useful for debugging or for initialising other dust objects.

Usage

dust_class$rng_state(first_only = FALSE)

Arguments

first_only

Logical, indicating if we should return only the first particle's random number state. If FALSE (the default) all particles states are returned, being 32 bytes per particle. If TRUE then we take just the first particle's state, which will be a total of 32 bytes. Both forms are suitable for seeding a new dust object as the shorter version will be used for the first particle, followed by jumps for each subsequent particle.

Method has_openmp()

Returns a logical, indicating if this model was compiled with "OpenMP" support, in which case it will react to the n_threads argument passed to the constructor. This method can also be used as a static method by running it directly as dust_class$public_methods$has_openmp()

Usage

dust_class$has_openmp()

Examples

# dust includes a couple of very simple examples
filename <- system.file("examples/walk.cpp", package = "dust")

# This model implements a random walk with a parameter coming from
# R representing the standard deviation of the walk
writeLines(readLines(filename))

#> class walk {
#> public:
#>   typedef double real_t;
#>   struct init_t {
#>     real_t sd;
#>   };
#>   walk(const init_t& data) : data_(data) {
#>   }
#>   size_t size() const {
#>     return 1;
#>   }
#>   std::vector<real_t> initial(size_t step) {
#>     std::vector<real_t> ret = {0};
#>     return ret;
#>   }
#>   void update(size_t step, const real_t * state,
#>               dust::rng_state_t<real_t>& rng_state,
#>               real_t * state_next) {
#>     real_t mean = state[0];
#>     state_next[0] = dust::distr::rnorm(rng_state, mean, data_.sd);
#>   }
#> 
#> private:
#>   init_t data_;
#> };
#> 
#> #include <cpp11/list.hpp>
#> template <>
#> walk::init_t dust_data<walk>(cpp11::list data) {
#>   walk::real_t sd = cpp11::as_cpp<walk::real_t>(data["sd"]);
#>   return walk::init_t{sd};
#> }

# The model can be compiled and loaded with dust::dust(filename)
# but it's faster in this example to use the prebuilt version in
# the package
model <- dust::dust_example("walk")

# Print the object and you can see the methods that it provides
model

#> <dust> object generator
#>   Public:
#>     initialize: function (data, step, n_particles, n_threads = 1L, seed = NULL) 
#>     name: function () 
#>     run: function (step_end) 
#>     set_index: function (index) 
#>     index: function () 
#>     n_threads: function () 
#>     set_state: function (state, step = NULL) 
#>     reset: function (data, step) 
#>     state: function (index = NULL) 
#>     step: function () 
#>     reorder: function (index) 
#>     info: function () 
#>     data: function () 
#>     rng_state: function (first_only = FALSE) 
#>     has_openmp: function () 
#>   Private:
#>     data_: NULL
#>     index_: NULL
#>     info_: NULL
#>     n_threads_: NULL
#>     ptr_: NULL
#>     simulate: function (steps, data, state, index = NULL, n_threads = 1L, seed = NULL) 
#>   Parent env: <environment: namespace:dust>
#>   Locked objects: TRUE
#>   Locked class: FALSE
#>   Portable: TRUE

# Create a model with standard deviation of 1, initial step zero
# and 30 particles
obj <- model$new(list(sd = 1), 0, 30)
obj

#> <dust>
#>   Public:
#>     data: function () 
#>     has_openmp: function () 
#>     index: function () 
#>     info: function () 
#>     initialize: function (data, step, n_particles, n_threads = 1L, seed = NULL) 
#>     n_threads: function () 
#>     name: function () 
#>     reorder: function (index) 
#>     reset: function (data, step) 
#>     rng_state: function (first_only = FALSE) 
#>     run: function (step_end) 
#>     set_index: function (index) 
#>     set_state: function (state, step = NULL) 
#>     state: function (index = NULL) 
#>     step: function () 
#>   Private:
#>     data_: list
#>     index_: NULL
#>     info_: NULL
#>     n_threads_: 1
#>     ptr_: externalptr
#>     simulate: function (steps, data, state, index = NULL, n_threads = 1L, seed = NULL) 

# Curent state is all zero
obj$state()

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,27] [,28] [,29] [,30]
#> [1,]     0     0     0     0

# Current step is also zero
obj$step()

#> [1] 0

# Run the model up to step 100
obj$run(100)

#>           [,1]      [,2]     [,3]      [,4]     [,5]     [,6]     [,7]
#> [1,] -8.629419 -9.102403 6.102584 0.9757289 16.83793 12.71314 -3.93674
#>           [,8]      [,9]     [,10]     [,11]     [,12]    [,13]   [,14]
#> [1,] -23.10724 -24.96031 0.5873703 -3.271908 -1.131626 11.91446 11.8673
#>          [,15]     [,16]      [,17]     [,18]    [,19]     [,20]    [,21]
#> [1,] -7.305056 0.2544087 -0.8322601 -11.50617 7.220508 -8.321633 2.843871
#>           [,22]    [,23]     [,24]     [,25]    [,26]     [,27]    [,28]
#> [1,] -0.7409265 9.872538 -1.538154 -6.480785 8.278559 -5.560665 5.151377
#>          [,29]    [,30]
#> [1,] -10.59943 -16.4027

# Reorder/resample the particles:
obj$reorder(sample(30, replace = TRUE))

# See the state again
obj$state()

#>          [,1]      [,2]     [,3]     [,4]      [,5]      [,6]      [,7]
#> [1,] 9.872538 -1.131626 16.83793 12.71314 -7.305056 -7.305056 0.9757289
#>          [,8]      [,9]    [,10]    [,11]     [,12]     [,13]    [,14]
#> [1,] 12.71314 -24.96031 16.83793 16.83793 -1.538154 -7.305056 11.91446
#>          [,15]    [,16]    [,17]     [,18]    [,19]     [,20]    [,21]    [,22]
#> [1,] -9.102403 -16.4027 9.872538 -3.271908 -16.4027 -3.271908 16.83793 2.843871
#>          [,23]     [,24]     [,25]    [,26]     [,27]     [,28]     [,29]
#> [1,] -1.131626 -10.59943 -9.102403 12.71314 0.9757289 0.2544087 -9.102403
#>         [,30]
#> [1,] 6.102584