Dust Random Number Generator

Create an object that can be used to generate random numbers with the same RNG as dust uses internally. This is primarily meant for debugging and testing.

Methods

Public methods

dust_rng$new()

dust_rng$size()

dust_rng$jump()

dust_rng$long_jump()

dust_rng$unif_rand()

dust_rng$norm_rand()

dust_rng$runif()

dust_rng$rnorm()

dust_rng$rbinom()

dust_rng$rpois()

dust_rng$rexp()

dust_rng$state()

Method new()

Create a dust_rng object

Usage

dust_rng$new(seed, n_generators = 1L)

Arguments

seed

The seed, as an integer or as a raw vector.

n_generators

The number of generators to use. While this function never runs in parallel, this is used to create a set of interleaved independent generators as dust would use in a model.

Method size()

Number of generators available

Usage

dust_rng$size()

Method jump()

The jump function for the generator, equivalent to 2^128 numbers drawn from the generator.

Usage

dust_rng$jump()

Method long_jump()

The long_jump function for the generator, equivalent to 2^192 numbers drawn from the generator. Generate n numbers from a standard uniform distribution

Usage

dust_rng$long_jump()

Method unif_rand()

Usage

dust_rng$unif_rand(n)

Arguments

n

Number of samples to draw Generate n numbers from a standard normal distribution

Method norm_rand()

Usage

dust_rng$norm_rand(n)

Arguments

n

Number of samples to draw Generate n numbers from a uniform distribution

Method runif()

Usage

dust_rng$runif(n, min, max)

Arguments

n

Number of samples to draw

min

The minimum of the distribution (length 1 or n)

max

The maximum of the distribution (length 1 or n) Generate n numbers from a normal distribution

Method rnorm()

Usage

dust_rng$rnorm(n, mean, sd)

Arguments

n

Number of samples to draw

mean

The mean of the distribution (length 1 or n)

sd

The standard deviation of the distribution (length 1 or n) Generate n numbers from a binomial distribution

Method rbinom()

Usage

dust_rng$rbinom(n, size, prob)

Arguments

n

Number of samples to draw

size

The number of trials (zero or more, length 1 or n)

prob

The probability of success on each trial (between 0 and 1, length 1 or n) Generate n numbers from a Poisson distribution

Method rpois()

Usage

dust_rng$rpois(n, lambda)

Arguments

n

Number of samples to draw

lambda

The mean (zero or more, length 1 or n) Generate n numbers from a exponential distribution

Method rexp()

Usage

dust_rng$rexp(n, rate)

Arguments

n

Number of samples to draw

rate

The rate of the exponential

Method state()

Returns the state of the random number generator. This returns a raw vector of length 32 * n_generators. It is primarily intended for debugging as one cannot (yet) initialise a dust_rng object with this state.

Usage

dust_rng$state()

Examples

rng <- dust_rng$new(42)

# Shorthand for Uniform(0, 1)
rng$unif_rand(5)

#> [1] 0.3614713 0.8809038 0.6228497 0.9013818 0.1483888

# Shorthand for Normal(0, 1)
rng$norm_rand(5)

#> [1] -0.8431672  0.6774323 -0.3107898  0.1477770  0.7396579

# Uniform random numbers between min and max
rng$runif(5, -2, 6)

#> [1]  2.607481  4.946376 -1.320962  4.695593  2.196486

# Normally distributed random numbers with mean and sd
rng$rnorm(5, 4, 2)

#> [1] 0.9321614 6.1411322 6.1099544 4.9142178 5.1787420

# Binomially distributed random numbers with size and prob
rng$rbinom(5, 10L, 0.3)

#> [1] 5 3 2 1 3

# Poisson distributed random numbers with mean lambda
rng$rpois(5, 2)

#> [1] 2 2 1 3 3