Advance a dust random number state

Advance a saved random number state by performing a "long jump" on it. If you have serialised the state using the $rng_state() method of a dust object but want create a new seed that is uncorrelated. If seed is extracted with $rng_seed() is to be reused multiple times, or if it will be used and the source object will also be used, then the state needs jumping to prevent generating the same sequence of random numbers.

dust_rng_state_long_jump(state, times = 1L)

Arguments

state
A raw vector representing dust random number generator; see dust_rng.

times
An integer indicating the number of times the long_jump should be performed. The default is one, but values larger than one will repeatedly advance the state.

Examples

# Create a new RNG object
rng <- dust::dust_rng$new(1)

# Serialise the state as a raw vector
state <- rng$state()

# We can advance this state
dust_rng_state_long_jump(state)

#>  [1] 77 69 c2 44 65 7b dc 93 49 c7 93 ef dd 06 fc 47 cf af f3 d0 80 97 01 68 45
#> [26] a2 90 f4 38 dd 4d 0a

# Which gives the same result as long_jump on the original generator
rng$long_jump()$state()

#>  [1] 77 69 c2 44 65 7b dc 93 49 c7 93 ef dd 06 fc 47 cf af f3 d0 80 97 01 68 45
#> [26] a2 90 f4 38 dd 4d 0a

rng$long_jump()$state()

#>  [1] 28 e8 de 5c 48 69 5b c3 44 02 9b 76 47 bd 81 a1 77 f7 63 47 53 29 1c e2 3a
#> [26] 0e 35 97 28 0a 59 c4

# Multiple jumps can be taken by using the "times" argument
dust_rng_state_long_jump(state, 2)

#>  [1] 28 e8 de 5c 48 69 5b c3 44 02 9b 76 47 bd 81 a1 77 f7 63 47 53 29 1c e2 3a
#> [26] 0e 35 97 28 0a 59 c4