Create a dust model from a C++ input file

Create a dust model from a C++ input file. This function will compile the dust support around your model and return an object that can be used to work with the model (see the Details below, and dust_class).

dust(filename, quiet = FALSE, workdir = NULL, gpu = FALSE, real_t = NULL)

Arguments

filename
The path to a single C++ file

quiet
Logical, indicating if compilation messages from pkgbuild should be displayed. Error messages will be displayed on compilation failure regardless of the value used.

workdir
Optional working directory to use. If NULL uses a temporary directory. By using a different directory of your choosing you can see the generated code.

gpu
Logical, indicating if we should generate GPU code. This requires a considerable amount of additional software installed (CUDA toolkit and drivers) as well as a CUDA-compatible GPU. If TRUE, then we call dust_cuda_options with no arguments. Alternatively, call that function and pass the value here (e.g, gpu = dust::dust_cuda_options(debug = TRUE)). Note that due to the use of the __syncwarp() primitive this may require a GPU with compute version 70 or higher.

real_t
Optionally, a string indicating a substitute type to swap in for your model's real_t declaration. If given, then we replace the string typedef (double|float) real_t with the given type. This is primarily intended to be used as gpu = TRUE, real_t = "float" in order to create model for the GPU that will use 32 bit floats (rather than 64 bit doubles, which are much slower). For CPU models decreasing precision of your real type will typically just decrease precision for no additional performance.

Input requirements

Your input dust model must satisfy a few requirements.

Define some class that implements your model (below model is assumed to be the class name)

That class must define a type internal_t (so model::internal_t) that contains its internal data that the model may change during execution (i.e., that is not shared between particles). If no such data is needed, you can do typedef dust::no_internal internal_t to indicate this.

We also need a type shared_t that contains constant internal data is shared between particles (e.g., dimensions, arrays that are read but not written). If no such data is needed, you can do typedef dust::no_shared shared_t to indicate this.

That class must also include a typedef that describes the model's floating point type, real_t. Most models can include typedef double real_t; in their public section.

The class must also include a typedef that describes the model's data type. This interface is subject to change, and for now you should include typedef dust::no_data data_t which marks your class as not supporting data, which disables the compare_data and set_data methods.

The class must have a constructor that accepts const dust::pars_t<model>& pars for your type model. This will have elements shared and internal which you can assign into your model if needed.

The model must have a method size() returning size_t which returns the size of the system. This size may depend on values in your initialisation object but is constant within a model run.

The model must have a method initial (which may not be const), taking a step number (size_t) and returning a std::vector<real_t> of initial state for the model.

The model must have a method update taking arguments:

size_t step: the step number

const double * state: the state at the beginning of the step

dust::rng_state_t<real_t>& rng_state: the dust random number generator state - this must be a reference, as it will be modified as random numbers are drawn

double *state_next: the end state of the model (to be written to by your function)

Your update function is the core here and should update the state of the system - you are expected to update all values of state on return.

It is very important that none of the functions in the class use the R API in any way as these functions will be called in parallel.

You must also provide a data/parameter-wrangling function for producing an object of type dust::pars_t<model> from an R list. We use cpp11 for this. Your function will look like:

namespace dust {
template <>
dust::pars_t<model> dust_pars<model>(cpp11::list pars) {
  // ...
  return dust::pars_t<model>(shared, internal);
}
}

With the body interacting with pars to create an object of type model::shared_t and model::internal_t before returning the dust::pars_t object. This function will be called in serial and may use anything in the cpp11 API. All elements of the returned object must be standard C/C++ (e.g., STL) types and not cpp11/R types. If your model uses only shared or internal, you may use the single-argument constructor overload to dust::pars_t which is equivalent to using dust::no_shared or dust::no_internal for the missing argument.

Your model may provided a template specialisation dust::dust_info<model>() returning a cpp11::sexp for returning arbitrary information back to the R session:

namespace dust {
template <>
cpp11::sexp dust_info<model>(const dust::pars_t<sir>& pars) {
  return cpp11::wrap(...);
}
}

What you do with this is up to you. If not present then the info() method on the created object will return NULL. Potential use cases for this are to return information about variable ordering, or any processing done while accepting the pars object used to create the pars fed into the particles.

Configuring your model

You can optionally use C++ pseudo-attributes to configure the generated code. Currently we support two attributes:

[[dust::class(classname)]] will tell dust the name of your target C++ class (in this example classname). You will need to use this if your file uses more than a single class, as otherwise will try to detect this using extremely simple heuristics.

[[dust::name(modelname)]] will tell dust the name to use for the class in R code. For technical reasons this must be alphanumeric characters only (sorry, no underscore) and must not start with a number. If not included then the C++ type name will be used (either specified with [[dust::class()]] or detected).

Error handling

Your model throw exceptions as a last resort. One such last resort exists if rbinom is given invalid inputs to prevent an infinite loop. If an error is thrown, all particles will complete their current run, and then the error will be rethrown

this is required by our parallel processing design. Once this happens though the state of the system is "inconsistent" as it contains particles that have run for different lengths of time. You can extract the state of the system at the point of failure (which may help with debugging) but you will be unable to continue running the object until either you reset it (with $reset()) or set both the state and step with $set_state(). An error will be thrown otherwise.

Things are worse on a GPU; if an error is thrown by the RNG code (happens in rbinom when given impossible inputs such as negative sizes, probabilities less than zero or greater than 1) then we currently use CUDA's __trap() function which will require a process restart to be able to use any device function again, covering all methods in the class. However, this is preferable to the infinite loop that would otherwise be caused.

See also

Examples

# dust includes a couple of very simple examples
filename <- system.file("examples/walk.cpp", package = "dust")

# This model implements a random walk with a parameter coming from
# R representing the standard deviation of the walk
writeLines(readLines(filename))

#> class walk {
#> public:
#>   typedef double real_t;
#>   typedef dust::no_data data_t;
#>   typedef dust::no_internal internal_t;
#> 
#>   struct shared_t {
#>     real_t sd;
#>   };
#> 
#>   walk(const dust::pars_t<walk>& pars) : shared(pars.shared) {
#>   }
#> 
#>   size_t size() const {
#>     return 1;
#>   }
#> 
#>   std::vector<real_t> initial(size_t step) {
#>     std::vector<real_t> ret = {0};
#>     return ret;
#>   }
#> 
#>   void update(size_t step, const real_t * state,
#>               dust::rng_state_t<real_t>& rng_state,
#>               real_t * state_next) {
#>     real_t mean = state[0];
#>     state_next[0] = dust::distr::rnorm(rng_state, mean, shared->sd);
#>   }
#> 
#> private:
#>   dust::shared_ptr<walk> shared;
#> };
#> 
#> namespace dust {
#> 
#> template <>
#> dust::pars_t<walk> dust_pars<walk>(cpp11::list pars) {
#>   walk::real_t sd = cpp11::as_cpp<walk::real_t>(pars["sd"]);
#>   return dust::pars_t<walk>(walk::shared_t{sd});
#> }
#> 
#> }

# The model can be compiled and loaded with dust::dust(filename)
# but it's faster in this example to use the prebuilt version in
# the package
model <- dust::dust_example("walk")

# Print the object and you can see the methods that it provides
model

#> <dust> object generator
#>   Public:
#>     initialize: function (pars, step, n_particles, n_threads = 1L, seed = NULL, 
#>     name: function () 
#>     param: function () 
#>     run: function (step_end, device = FALSE) 
#>     simulate: function (step_end, device = FALSE) 
#>     set_index: function (index) 
#>     index: function () 
#>     n_threads: function () 
#>     n_state: function () 
#>     n_particles: function () 
#>     n_particles_each: function () 
#>     shape: function () 
#>     set_state: function (state, step = NULL) 
#>     reset: function (pars, step) 
#>     set_pars: function (pars) 
#>     state: function (index = NULL) 
#>     step: function () 
#>     reorder: function (index) 
#>     resample: function (weights) 
#>     info: function () 
#>     pars: function () 
#>     rng_state: function (first_only = FALSE, last_only = FALSE) 
#>     set_rng_state: function (rng_state) 
#>     has_openmp: function () 
#>     has_cuda: function () 
#>     n_pars: function () 
#>     set_n_threads: function (n_threads) 
#>     has_compare: function () 
#>     set_data: function (data) 
#>     compare_data: function (device = FALSE) 
#>     filter: function (save_trajectories = FALSE, step_snapshot = NULL, device = FALSE) 
#>     device_info: function () 
#>   Private:
#>     pars_: NULL
#>     pars_multi_: NULL
#>     index_: NULL
#>     info_: NULL
#>     n_threads_: NULL
#>     n_particles_: NULL
#>     n_particles_each_: NULL
#>     shape_: NULL
#>     ptr_: NULL
#>     device_id_: NULL
#>     param_: NULL
#>   Parent env: <environment: namespace:dust>
#>   Locked objects: TRUE
#>   Locked class: FALSE
#>   Portable: TRUE

# Create a model with standard deviation of 1, initial step zero
# and 30 particles
obj <- model$new(list(sd = 1), 0, 30)
obj

#> <dust>
#>   Public:
#>     compare_data: function (device = FALSE) 
#>     device_info: function () 
#>     filter: function (save_trajectories = FALSE, step_snapshot = NULL, device = FALSE) 
#>     has_compare: function () 
#>     has_cuda: function () 
#>     has_openmp: function () 
#>     index: function () 
#>     info: function () 
#>     initialize: function (pars, step, n_particles, n_threads = 1L, seed = NULL, 
#>     n_pars: function () 
#>     n_particles: function () 
#>     n_particles_each: function () 
#>     n_state: function () 
#>     n_threads: function () 
#>     name: function () 
#>     param: function () 
#>     pars: function () 
#>     reorder: function (index) 
#>     resample: function (weights) 
#>     reset: function (pars, step) 
#>     rng_state: function (first_only = FALSE, last_only = FALSE) 
#>     run: function (step_end, device = FALSE) 
#>     set_data: function (data) 
#>     set_index: function (index) 
#>     set_n_threads: function (n_threads) 
#>     set_pars: function (pars) 
#>     set_rng_state: function (rng_state) 
#>     set_state: function (state, step = NULL) 
#>     shape: function () 
#>     simulate: function (step_end, device = FALSE) 
#>     state: function (index = NULL) 
#>     step: function () 
#>   Private:
#>     device_id_: -1
#>     index_: NULL
#>     info_: NULL
#>     n_particles_: 30
#>     n_particles_each_: 30
#>     n_threads_: 1
#>     param_: NULL
#>     pars_: list
#>     pars_multi_: FALSE
#>     ptr_: externalptr
#>     shape_: 30

# Curent state is all zero
obj$state()

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,27] [,28] [,29] [,30]
#> [1,]     0     0     0     0

# Current step is also zero
obj$step()

#> [1] 0

# Run the model up to step 100
obj$run(100)

#>           [,1]      [,2]     [,3]      [,4]     [,5]     [,6]     [,7]
#> [1,] -8.629419 -9.102403 6.102584 0.9757289 16.83793 12.71314 -3.93674
#>           [,8]      [,9]     [,10]     [,11]     [,12]    [,13]   [,14]
#> [1,] -23.10724 -24.96031 0.5873703 -3.271908 -1.131626 11.91446 11.8673
#>          [,15]     [,16]      [,17]     [,18]    [,19]     [,20]    [,21]
#> [1,] -7.305056 0.2544087 -0.8322601 -11.50617 7.220508 -8.321633 2.843871
#>           [,22]    [,23]     [,24]     [,25]    [,26]     [,27]    [,28]
#> [1,] -0.7409265 9.872538 -1.538154 -6.480785 8.278559 -5.560665 5.151377
#>          [,29]    [,30]
#> [1,] -10.59943 -16.4027

# Reorder/resample the particles:
obj$reorder(sample(30, replace = TRUE))

# See the state again
obj$state()

#>          [,1]      [,2]     [,3]     [,4]      [,5]      [,6]      [,7]
#> [1,] 9.872538 -1.131626 16.83793 12.71314 -7.305056 -7.305056 0.9757289
#>          [,8]      [,9]    [,10]    [,11]     [,12]     [,13]    [,14]
#> [1,] 12.71314 -24.96031 16.83793 16.83793 -1.538154 -7.305056 11.91446
#>          [,15]    [,16]    [,17]     [,18]    [,19]     [,20]    [,21]    [,22]
#> [1,] -9.102403 -16.4027 9.872538 -3.271908 -16.4027 -3.271908 16.83793 2.843871
#>          [,23]     [,24]     [,25]    [,26]     [,27]     [,28]     [,29]
#> [1,] -1.131626 -10.59943 -9.102403 12.71314 0.9757289 0.2544087 -9.102403
#>         [,30]
#> [1,] 6.102584