Create a dust model from a C++ input file

Create a dust model from a C++ input file. This function will compile the dust support around your model and return an object that can be used to work with the model (see the Details below).

dust(filename, quiet = FALSE, workdir = NULL)

Arguments

filename
The path to a single C++ file

quiet
Logical, indicating if compilation messages from pkgbuild should be displayed. Error messages will be displayed on compilation failure regardless of the value used.

workdir
Optional working directory to use. If NULL uses a temporary directory. By using a different directory of your choosing you can see the generated code.

Input requirements

Your input dust model must satisfy a few requirements.

Define some class that implements your model (below model is assumed to be the class name)

That class must define a type internal_t (so model::internal_t) that contains its internal data that the model may change during execution (i.e., that is not shared between particles). If no such data is needed, you can do typedef dust::no_internal internal_t to indicate this.

We also need a type shared_t that contains constant internal data is shared between particles (e.g., dimensions, arrays that are read but not written). If no such data is needed, you can do typedef dust::no_shared shared_t to indicate this.

That class must also include a typedef that describes the model's floating point type, real_t. Most models can include typedef double real_t; in their public section.

The class must also include a typedef that describes the model's data type. This interface is subject to change, and for now you should include typedef dust::no_data data_t which marks your class as not supporting data, which disables the compare_data and set_data methods.

The class must have a constructor that accepts const dust::pars_t<model>& pars for your type model. This will have elements shared and internal which you can assign into your model if needed.

The model must have a method size() returning size_t which returns the size of the system. This size may depend on values in your initialisation object but is constant within a model run.

The model must have a method initial (which may not be const), taking a step number (size_t) and returning a std::vector<real_t> of initial state for the model.

The model must have a method update taking arguments:

size_t step: the step number

const double * state: the state at the beginning of the step

dust::rng_state_t<real_t>& rng_state: the dust random number generator state - this must be a reference, as it will be modified as random numbers are drawn

double *state_next: the end state of the model (to be written to by your function)

Your update function is the core here and should update the state of the system - you are expected to update all values of state on return.

It is very important that none of the functions in the class use the R API in any way as these functions will be called in parallel.

You must also provide a data/parameter-wrangling function for producing an object of type dust::pars_t<model> from an R list. We use cpp11 for this. Your function will look like:

template <>
dust::pars_t<model> dust_pars<model>(cpp11::list pars) {
  // ...
  return dust::pars_t<model>(shared, internal);
}

With the body interacting with pars to create an object of type model::shared_t and model::internal_t before returning the dust::pars_t object. This function will be called in serial and may use anything in the cpp11 API. All elements of the returned object must be standard C/C++ (e.g., STL) types and not cpp11/R types. If your model uses only shared or internal, you may use the single-argument constructor overload to dust::pars_t which is equivalent to using dust::no_shared or dust::no_internal for the missing argument.

Your model may provided a template specialisation dust_info<model>() returning a cpp11::sexp for returning arbitrary information back to the R session:

template <>
cpp11::sexp dust_info<model>(const dust::pars_t<sir>& pars) {
  return cpp11::wrap(...);
}

What you do with this is up to you. If not present then the info() method on the created object will return NULL. Potential use cases for this are to return information about variable ordering, or any processing done while accepting the pars object used to create the pars fed into the particles.

Configuring your model

You can optionally use C++ psuedo-attributes to configure the generated code. Currently we support two attributes:

[[dust::class(classname)]] will tell dust the name of your target C++ class (in this example classname). You will need to use this if your file uses more than a single class, as otherwise will try to detect this using extremely simple heuristics.

[[dust::name(modelname)]] will tell dust the name to use for the class in R code. For technical reasons this must be alphanumeric characters only (sorry, no underscore) and must not start with a number. If not included then the C++ type name will be used (either specified with [[dust::class()]] or detected).

Methods

Public methods

dust_class$new()

dust_class$name()

dust_class$param()

dust_class$run()

dust_class$set_index()

dust_class$index()

dust_class$n_threads()

dust_class$n_state()

dust_class$n_particles()

dust_class$set_state()

dust_class$reset()

dust_class$set_pars()

dust_class$state()

dust_class$step()

dust_class$reorder()

dust_class$resample()

dust_class$info()

dust_class$pars()

dust_class$rng_state()

dust_class$set_rng_state()

dust_class$has_openmp()

dust_class$n_pars()

dust_class$set_n_threads()

dust_class$has_compare()

dust_class$set_data()

dust_class$compare_data()

Method new()

Create a new model. Note that the behaviour of this object created by this function will change considerably based on whether the pars_multi argument is TRUE. If not (the default) then we create n_particles which all share the same parameters as specified by the pars argument. If pars_multi is TRUE then pars must be an unnamed list, and each element of it represents a different set of parameters. We will create length(pars) sets of n_particles particles which will be simulated together. These particles must have the same dimension - that is, they must correspond to model state that is the same size.

Usage

dust_class$new(
  pars,
  step,
  n_particles,
  n_threads = 1L,
  seed = NULL,
  pars_multi = FALSE
)

Arguments

pars

Data to initialise your model with; a list object, but the required elements will depend on the details of your model. If pars_multi is TRUE, then this must be an unnamed list of pars objects (see Details).

step

Initial step - must be nonnegative

n_particles

Number of particles to create - must be at least 1

n_threads

Number of OMP threads to use, if dust and your model were compiled with OMP support (details to come). n_particles should be a multiple of n_threads (e.g., if you use 8 threads, then you should have 8, 16, 24, etc particles). However, this is not compulsary.

seed

The seed to use for the random number generator. Can be a positive integer, NULL (initialise with R's random number generator) or a raw vector of a length that is a multiple of 32 to directly initialise the generator (e..g., from the dust object's $rng_state() method).

pars_multi

Logical, indicating if pars should be interpreted as a set of different initialisations, and that we should prepare n_particles * length(pars) particles for simulation. This has an effect on many of the other methods of the object.

Method name()

Returns friendly model name

Usage

dust_class$name()

Method param()

Returns parameter information, if provided by the model. This describes the contents of pars passed to the constructor or to reset as the pars argument, and the details depend on the model.

Usage

dust_class$param()

Method run()

Run the model up to a point in time, returning the filtered state at that point.

Usage

dust_class$run(step_end)

Arguments

step_end

Step to run to (if less than or equal to the current step(), silently nothing will happen)

Method set_index()

Set the "index" vector that is used to return a subset of pars after using run(). If this is not used then run() returns all elements in your state vector, which may be excessive and slower than necessary. This method must be called after any call to reset() as reset() may change the size of the state and that will invalidate the index.

Usage

dust_class$set_index(index)

Arguments

index

The index vector - must be an integer vector with elements between 1 and the length of the state (this will be validated, and an error thrown if an invalid index is given).

Method index()

Returns the index as set by $set_index

Usage

dust_class$index()

Method n_threads()

Returns the number of threads that the model was constructed with

Usage

dust_class$n_threads()

Method n_state()

Returns the length of the per-particle state

Usage

dust_class$n_state()

Method n_particles()

Returns the number of particles

Usage

dust_class$n_particles()

Method set_state()

Set the "state" vector for all particles, overriding whatever your models initial() method provides.

Usage

dust_class$set_state(state, step = NULL)

Arguments

state

The state vector - can be either a numeric vector with the same length as the model's current state (in which case the same state is applied to all particles), or a numeric matrix with as many rows as your model's state and as many columns as you have particles (in which case you can set a number of different starting states at once).

step

If not NULL, then this sets the initial step. If this is a vector (with the same length as the number of particles), then particles are started from different initial steps and run up to the larges step given (i.e., max(step))

Method reset()

Reset the model while preserving the random number stream state

Usage

dust_class$reset(pars, step)

Arguments

pars

New pars for the model (see constructor)

step

New initial step for the model (see constructor)

Method set_pars()

Set the 'pars' element in a dust object while holding model state, index, etc constant. In contrast to $reset, the old state must be compatible with the new one (e.g., don't change model size), and the index will remain valid.

Usage

dust_class$set_pars(pars)

Arguments

pars

New pars for the model (see constructor)

Method state()

Return full model state

Usage

dust_class$state(index = NULL)

Arguments

index

Optional index to select state using

Method step()

Return current model step

Usage

dust_class$step()

Method reorder()

Reorder particles.

Usage

dust_class$reorder(index)

Arguments

index

An integer vector, with values between 1 and n_particles, indicating the index of the current particles that new particles should take.

Method resample()

Reasmple particles according to some weight.

Usage

dust_class$resample(weights)

Arguments

weights

A numeric vector representing particle weights. For a "multi-parameter" dust object this should be be a matrix with the number of rows being the number of particles per parameter set and the number of columns being the number of parameter sets. long as all particles or be a matrix.

Method info()

Returns information about the pars that your model was created with. Only returns non-NULL if the model provides a dust_info template specialisation.

Usage

dust_class$info()

Method pars()

Returns the pars object that your model was constructed with.

Usage

dust_class$pars()

Method rng_state()

Returns the state of the random number generator. This returns a raw vector of length 32 * n_particles. This can be useful for debugging or for initialising other dust objects.

Usage

dust_class$rng_state(first_only = FALSE)

Arguments

first_only

Logical, indicating if we should return only the first particle's random number state. If FALSE (the default) all particles states are returned, being 32 bytes per particle. If TRUE then we take just the first particle's state, which will be a total of 32 bytes. Both forms are suitable for seeding a new dust object as the shorter version will be used for the first particle, followed by jumps for each subsequent particle.

Method set_rng_state()

Set the random number state for this model. This replaces the RNG state that the model is using with a state of your choosing, saved out from a different model object. This method is designed to support advanced use cases where it is easier to manipulate the state of the random number generator than the internal state of the dust object.

Usage

dust_class$set_rng_state(rng_state)

Arguments

rng_state

A random number state, as saved out by the $rng_state() method. Note that unlike seed as passed to the constructor, this must be a raw vector of the expected length.

Method has_openmp()

Returns a logical, indicating if this model was compiled with "OpenMP" support, in which case it will react to the n_threads argument passed to the constructor. This method can also be used as a static method by running it directly as dust_class$public_methods$has_openmp()

Usage

dust_class$has_openmp()

Method n_pars()

Returns the number of distinct pars elements required. This is 0 where the object was initialised with pars_multi = FALSE and an integer otherwise. For multi-pars dust objects, Where pars is accepted, you must provide an unnamed list of length $n_pars().

Usage

dust_class$n_pars()

Method set_n_threads()

Change the number of threads that the dust object will use. Your model must be compiled with "OpenMP" support for this to have an effect. Returns (invisibly) the previous value.

Usage

dust_class$set_n_threads(n_threads)

Arguments

n_threads

The new number of threads to use. You may want to wrap this argument in dust_openmp_threads() in order to verify that you can actually use the number of threads requested (based on environment variables and OpenMP support).

Method has_compare()

Returns a logical, indicating if this model was compiled with "compare" support, in which case the set_data and compare_data methods are available (otherwise these methods will error). This method can also be used as a static method by running it directly as dust_class$public_methods$has_compare()

Usage

dust_class$has_compare()

Method set_data()

Set "data" into the model for use with the $compare_data() method. This is not supported by all models, depending on if they define a data_t type. See dust_data() for a helper function to construct suitable data and a description of the required format. You will probably want to use that here, and definitely if using multiple parameter sets.

Usage

dust_class$set_data(data)

Arguments

data

A list of data to set.

Method compare_data()

Compare the current model state against the data as set by set_data. If there is no data set, or no data corresponding to the current time then NULL is returned. Otherwise a numeric vector the same length as the number of particles is returned. If model's underlying compare_data function is stochastic, then each call to this function may be result in a different answer.

Usage

dust_class$compare_data()

Examples

# dust includes a couple of very simple examples
filename <- system.file("examples/walk.cpp", package = "dust")

# This model implements a random walk with a parameter coming from
# R representing the standard deviation of the walk
writeLines(readLines(filename))

#> class walk {
#> public:
#>   typedef double real_t;
#>   typedef dust::no_data data_t;
#>   typedef dust::no_internal internal_t;
#> 
#>   struct shared_t {
#>     real_t sd;
#>   };
#> 
#>   walk(const dust::pars_t<walk>& pars) : shared(pars.shared) {
#>   }
#> 
#>   size_t size() const {
#>     return 1;
#>   }
#> 
#>   std::vector<real_t> initial(size_t step) {
#>     std::vector<real_t> ret = {0};
#>     return ret;
#>   }
#> 
#>   void update(size_t step, const real_t * state,
#>               dust::rng_state_t<real_t>& rng_state,
#>               real_t * state_next) {
#>     real_t mean = state[0];
#>     state_next[0] = dust::distr::rnorm(rng_state, mean, shared->sd);
#>   }
#> 
#> private:
#>   dust::shared_ptr<walk> shared;
#> };
#> 
#> #include <cpp11/list.hpp>
#> template <>
#> dust::pars_t<walk> dust_pars<walk>(cpp11::list pars) {
#>   walk::real_t sd = cpp11::as_cpp<walk::real_t>(pars["sd"]);
#>   return dust::pars_t<walk>(walk::shared_t{sd});
#> }

# The model can be compiled and loaded with dust::dust(filename)
# but it's faster in this example to use the prebuilt version in
# the package
model <- dust::dust_example("walk")

# Print the object and you can see the methods that it provides
model

#> <dust> object generator
#>   Public:
#>     initialize: function (pars, step, n_particles, n_threads = 1L, seed = NULL, 
#>     name: function () 
#>     param: function () 
#>     run: function (step_end) 
#>     set_index: function (index) 
#>     index: function () 
#>     n_threads: function () 
#>     n_state: function () 
#>     n_particles: function () 
#>     set_state: function (state, step = NULL) 
#>     reset: function (pars, step) 
#>     set_pars: function (pars) 
#>     state: function (index = NULL) 
#>     step: function () 
#>     reorder: function (index) 
#>     resample: function (weights) 
#>     info: function () 
#>     pars: function () 
#>     rng_state: function (first_only = FALSE) 
#>     set_rng_state: function (rng_state) 
#>     has_openmp: function () 
#>     n_pars: function () 
#>     set_n_threads: function (n_threads) 
#>     has_compare: function () 
#>     set_data: function (data) 
#>     compare_data: function () 
#>   Private:
#>     pars_: NULL
#>     pars_multi_: NULL
#>     index_: NULL
#>     info_: NULL
#>     n_threads_: NULL
#>     n_particles_: NULL
#>     ptr_: NULL
#>     param_: NULL
#>     simulate: function (steps, pars, state, index = NULL, n_threads = 1L, seed = NULL, 
#>     check_pars_multi: function (pars, len) 
#>   Parent env: <environment: namespace:dust>
#>   Locked objects: TRUE
#>   Locked class: FALSE
#>   Portable: TRUE

# Create a model with standard deviation of 1, initial step zero
# and 30 particles
obj <- model$new(list(sd = 1), 0, 30)
obj

#> <dust>
#>   Public:
#>     compare_data: function () 
#>     has_compare: function () 
#>     has_openmp: function () 
#>     index: function () 
#>     info: function () 
#>     initialize: function (pars, step, n_particles, n_threads = 1L, seed = NULL, 
#>     n_pars: function () 
#>     n_particles: function () 
#>     n_state: function () 
#>     n_threads: function () 
#>     name: function () 
#>     param: function () 
#>     pars: function () 
#>     reorder: function (index) 
#>     resample: function (weights) 
#>     reset: function (pars, step) 
#>     rng_state: function (first_only = FALSE) 
#>     run: function (step_end) 
#>     set_data: function (data) 
#>     set_index: function (index) 
#>     set_n_threads: function (n_threads) 
#>     set_pars: function (pars) 
#>     set_rng_state: function (rng_state) 
#>     set_state: function (state, step = NULL) 
#>     state: function (index = NULL) 
#>     step: function () 
#>   Private:
#>     check_pars_multi: function (pars, len) 
#>     index_: NULL
#>     info_: NULL
#>     n_particles_: 30
#>     n_threads_: 1
#>     param_: NULL
#>     pars_: list
#>     pars_multi_: FALSE
#>     ptr_: externalptr
#>     simulate: function (steps, pars, state, index = NULL, n_threads = 1L, seed = NULL, 

# Curent state is all zero
obj$state()

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,27] [,28] [,29] [,30]
#> [1,]     0     0     0     0

# Current step is also zero
obj$step()

#> [1] 0

# Run the model up to step 100
obj$run(100)

#>           [,1]      [,2]     [,3]      [,4]     [,5]     [,6]     [,7]
#> [1,] -8.629419 -9.102403 6.102584 0.9757289 16.83793 12.71314 -3.93674
#>           [,8]      [,9]     [,10]     [,11]     [,12]    [,13]   [,14]
#> [1,] -23.10724 -24.96031 0.5873703 -3.271908 -1.131626 11.91446 11.8673
#>          [,15]     [,16]      [,17]     [,18]    [,19]     [,20]    [,21]
#> [1,] -7.305056 0.2544087 -0.8322601 -11.50617 7.220508 -8.321633 2.843871
#>           [,22]    [,23]     [,24]     [,25]    [,26]     [,27]    [,28]
#> [1,] -0.7409265 9.872538 -1.538154 -6.480785 8.278559 -5.560665 5.151377
#>          [,29]    [,30]
#> [1,] -10.59943 -16.4027

# Reorder/resample the particles:
obj$reorder(sample(30, replace = TRUE))

# See the state again
obj$state()

#>          [,1]      [,2]     [,3]     [,4]      [,5]      [,6]      [,7]
#> [1,] 9.872538 -1.131626 16.83793 12.71314 -7.305056 -7.305056 0.9757289
#>          [,8]      [,9]    [,10]    [,11]     [,12]     [,13]    [,14]
#> [1,] 12.71314 -24.96031 16.83793 16.83793 -1.538154 -7.305056 11.91446
#>          [,15]    [,16]    [,17]     [,18]    [,19]     [,20]    [,21]    [,22]
#> [1,] -9.102403 -16.4027 9.872538 -3.271908 -16.4027 -3.271908 16.83793 2.843871
#>          [,23]     [,24]     [,25]    [,26]     [,27]     [,28]     [,29]
#> [1,] -1.131626 -10.59943 -9.102403 12.71314 0.9757289 0.2544087 -9.102403
#>         [,30]
#> [1,] 6.102584